Services done

+++ src/main/java/io/zipcoder/persistenceapp/repository/DepartmentRepo.java
@@ -4,6 +4,9 @@ import io.zipcoder.persistenceapp.model.Department;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface DepartmentRepo extends CrudRepository<Department, Integer> {
+public interface DepartmentRepo extends CrudRepository<Department, Long> {
+
 }
