+++ src/main/java/io/zipcoder/persistenceapp/repository/EmployeeRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmployeeRepo extends CrudRepository<Employee, Integer> {
+public interface EmployeeRepo extends CrudRepository<Employee, Long> {
 }
